% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoAPAc.R
\name{AnnoAPAc}
\alias{AnnoAPAc}
\title{Annotate the APA clusters from a GTF file}
\usage{
AnnoAPAc(APAc_file, gtf_file, w_strand, add_chr = FALSE)
}
\arguments{
\item{APAc_file}{APA clusters file from MapAPAc()}

\item{gtf_file}{A reference GTF file}

\item{w_strand}{which strand, "+" or "-"}

\item{add_chr}{whether to add character "chr" on the chr_name of the GTF (default FALSE)}
}
\value{
Null, An annotated file in the same dir.
}
\description{
Annotate the APA clusters from a GTF file
}
\examples{
AnnoAPAc("APA_map.out", "hg19.gtf", "-")
}
