% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAPAc.R
\name{CountAPAc}
\alias{CountAPAc}
\title{Create a APA_cluster-Cell_Barcode count matrix for Seurart analysis}
\usage{
CountAPAc(
  cluster_p,
  cluster_n,
  in_bamfile,
  cb_file,
  chr_info,
  out_dir,
  filter_method = "end",
  UBtag = "UB"
)
}
\arguments{
\item{cluster_p}{Annotated APA clusters file of the positive strand, generated by AnnoAPAc()}

\item{cluster_n}{Annotated APA clusters file of the positive strand, generated by AnnoAPAc()}

\item{in_bamfile}{A indexed BAM file, and the index files (.bai) must be in the same directory of the indexed BAM}

\item{cb_file}{A cell barcodes list, can be generated by extractAPA() or cellranger}

\item{chr_info}{A data.table that contains the chrinfo, can be generated by ExtractChrinfo()}

\item{out_dir}{The directory where a directory named "matrix/" is created and the output files are stored}

\item{filter_method}{The way to  filter reads (defult "end")}

\item{UBtag}{UB tag (defult "UB")}
}
\value{
NULL. writes output in the directory "matrix/" of the out_dir
}
\description{
Create a APA_cluster-Cell_Barcode count matrix for Seurart analysis
}
\examples{
CountAPAc("p/APA_map.out.anno", "n/APA_map.out.anno", "test.bam", "_CBlist.txt", ExtractChrinfo("test.s.bam"), "output/")
}
