% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAPA.R
\name{ExtractAPA}
\alias{ExtractAPA}
\title{Generate strand-specific two-dimensional coordinate files (prebed) of the BAM for clustering, and extract the cb list}
\usage{
ExtractAPA(
  in_bamfile,
  in_nSBfile,
  chr_info,
  out_dir,
  SA_cutoff = 5,
  min_median = 5,
  RCn = 1
)
}
\arguments{
\item{in_bamfile}{A indexed BAM file, and the index files (.bai) must be in the same directory of the indexed BAM}

\item{in_nSBfile}{A indexed soft-clipped BAM file, generated and indexed by samtools script}

\item{chr_info}{A data.table that contains the chrinfo, can be generated by ExtractChrinfo()}

\item{out_dir}{The directory where directories of two strands are created and the output files are stored}

\item{SA_cutoff}{The minimum number of soft-clipped nts (default 5)}

\item{min_median}{The minimum number of locations of the same umi when screening (default 5)}

\item{RCn}{The cut-off value that decides the minimum weight of output when processing the in_bamfile (default 1)}
}
\value{
NULL, write the prebeds (sa, rc_end, rc_uni) in strand dirs under the out_dir and the cb list in out_dir.
}
\description{
Generate strand-specific two-dimensional coordinate files (prebed) of the BAM for clustering, and extract the cb list
}
\examples{
ExtractAPA("test.bam", "test.s.bam", ExtractChrinfo("test.bam"), "output/")
}
