% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAPAc.R
\name{FilterAPAc}
\alias{FilterAPAc}
\title{Filter internal APAc}
\usage{
FilterAPAc(
  APAc_file,
  fasta_file,
  input_strand,
  out_dir,
  dist = c(20, 50),
  AT_rich = "AAAAAAA"
)
}
\arguments{
\item{APAc_file}{APA clusters file from ClusterAPA()}

\item{fasta_file}{fasta file with an index}

\item{input_strand}{strand}

\item{out_dir}{The directory where the output files are stored}

\item{dist}{Scanning range (default c(20, 50))}

\item{AT_rich}{A rich motif (default "AAAAAAA")}
}
\value{
NULL, writes the output "APA_clusters.filtered.out" in the out_dir.
}
\description{
Filter internal APAc
}
\examples{
FilterAPAc("APA_clusters.out", "hg19.fa", "-", "output/")
}
