% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAPAc.R
\name{MapAPAc}
\alias{MapAPAc}
\title{Map APAc with prebed to count region}
\usage{
MapAPAc(
  RE_prebed,
  SA_prebed,
  APAc_file,
  out_dir,
  blocksize = 250,
  APAdist_cut = 25,
  SA_weight = 3
)
}
\arguments{
\item{RE_prebed}{RE file of one strand generated from ExtractAPA()}

\item{SA_prebed}{SA file of one strand generated from ExtractAPA()}

\item{APAc_file}{APA clusters file from FilterAPAc()}

\item{out_dir}{The directory where the output files are stored}

\item{blocksize}{Threshold for isolated nodes and clusters (default 250)}

\item{APAdist_cut}{Threshold for clustering the APA (default 25)}

\item{SA_weight}{The weight increase of SA file (default 3)}
}
\value{
NULL, writes the output "APA_map.out" in the out_dir.
}
\description{
Map APAc with prebed to count region
}
\examples{
MapAPAc("rc_end.prebed", "sa.prebed", "APA_clusters.filtered.out", "output/")
}
